microcom: initialize the struct sigaction before use
sigaction(2) reads the sa_mask and sa_flags struct members as well, but
we didn't initialize them so far. We probably didn't run into problems,
because we allocate it when the stack is fresh and full of zeroes.

Fix this by explicitly zeroing the struct and emptying the signal set.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>